refactor(ImageUpload): 将图片上传逻辑移动到扩展目录并删除旧hook
将useImgUpload hook从全局hooks目录移动到ImageUpload扩展目录
简化ImageUploader组件逻辑，移除重复的上传处理代码
